fix(storybook): Add story with OpenApiAdmin
