+++ api/config/packages/api_platform.yaml
@@ -6,6 +6,7 @@ api_platform:
         include_type: true
     formats:
         jsonld: ['application/ld+json']
+        json: ["application/json"]
     docs_formats:
         jsonld: ['application/ld+json']
         jsonopenapi: ['application/vnd.openapi+json']
