--- src/stories/Configure.mdx
@@ -1,3 +1,0 @@
-import { Meta } from "@storybook/blocks";
-
-<Meta title="Configure your project" />
